`@remotion/canvas`: Add headless timeline primitives
