+++ packages/canvas/README.md
@@ -1,0 +1,6 @@
+# @remotion/canvas
+
+Headless primitives for Remotion authoring interfaces.
+
+This package is an internal dependency of Remotion Studio. Its API is not yet
+ready for general use.
