+++ packages/canvas/src/get-connected-compositions.ts
@@ -1,0 +1,20 @@
+export const getConnectedCompositions = <
+	Composition extends {readonly componentFromProps?: unknown},
+>({
+	compositions,
+	singleChildComponent,
+}: {
+	readonly compositions: readonly Composition[];
+	readonly singleChildComponent: unknown;
+}): Composition[] => {
+	if (
+		singleChildComponent === null ||
+		typeof singleChildComponent === 'undefined'
+	) {
+		return [];
+	}
+
+	return compositions.filter(
+		(composition) => composition.componentFromProps === singleChildComponent,
+	);
+};
