+++ packages/canvas/src/timeline-sequence-node-path-to-key.ts
@@ -1,0 +1,5 @@
+import type {SequencePropsSubscriptionKey} from 'remotion';
+
+export const timelineSequenceNodePathToKey = (
+	key: SequencePropsSubscriptionKey,
+): string => `${key.absolutePath}:${JSON.stringify(key.nodePath)}`;
