+++ packages/canvas/tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"extends": "../tsconfig.settings.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"outDir": "dist",
+		"skipLibCheck": true,
+		"target": "ES2022",
+		"module": "es2020",
+		"moduleResolution": "bundler",
+		"declaration": true,
+		"emitDeclarationOnly": true
+	},
+	"include": ["src"],
+	"references": [{"path": "../core"}]
+}
