+++ packages/studio-shared/src/package-info.ts
@@ -11,6 +11,7 @@ const allPackages = [
 	'brand',
 	'bundler',
 	'browser-studio',
+	'canvas',
 	'claude-code-plugin',
 	'cli',
 	'cloudrun',
@@ -170,6 +171,7 @@ export const descriptions: {[key in Pkgs]: string | null} = {
 	lambda: 'Render Remotion videos on AWS Lambda',
 	bundler: 'Bundle Remotion compositions using Webpack',
 	'browser-studio': 'Run Remotion Studio in the browser',
+	canvas: 'Headless primitives for Remotion authoring interfaces',
 	'studio-codemods': 'Shared codemods for Remotion Studio',
 	'studio-server': 'Run a Remotion Studio with a server backend',
 	'install-whisper-cpp': 'Helpers for installing and using Whisper.cpp',
@@ -279,6 +281,7 @@ export const installableMap: {[key in Pkgs]: boolean} = {
 	brand: false,
 	bundler: false,
 	'browser-studio': false,
+	canvas: false,
 	cli: false,
 	cloudrun: true,
 	'claude-code-plugin': false,
@@ -403,6 +406,7 @@ export const apiDocs: {[key in Pkgs]: string | null} = {
 	lambda: 'https://www.remotion.dev/docs/lambda',
 	bundler: 'https://www.remotion.dev/docs/bundler',
 	'browser-studio': null,
+	canvas: null,
 	'studio-codemods': null,
 	'lambda-client': null,
 	'serverless-client': null,
