+++ packages/studio/src/helpers/get-sequence-visible-range.ts
@@ -1,75 +1,6 @@
-import type {TSequence} from 'remotion';
-
-export const getCascadedStart = (
-	sequence: TSequence,
-	sequences: TSequence[],
-): number => {
-	if (!sequence.parent) {
-		return sequence.from;
-	}
-
-	const parent = sequences.find((s) => s.id === sequence.parent);
-	if (!parent) {
-		throw new TypeError('Parent not found for sequence ' + sequence.id);
-	}
-
-	return getCascadedStart(parent, sequences) + sequence.from;
-};
-
-export const getCascadedStartWithTrim = (
-	sequence: TSequence,
-	sequences: TSequence[],
-): number => {
-	const effectiveFrom =
-		sequence.trimBefore === null
-			? sequence.from
-			: sequence.from - sequence.trimBefore;
-	if (!sequence.parent) {
-		return effectiveFrom;
-	}
-
-	const parent = sequences.find((s) => s.id === sequence.parent);
-	if (!parent) {
-		throw new TypeError('Parent not found for sequence ' + sequence.id);
-	}
-
-	return getCascadedStartWithTrim(parent, sequences) + effectiveFrom;
-};
-
-export const getTimelineVisibleStart = (
-	sequence: TSequence,
-	sequences: TSequence[],
-): number => {
-	const cascadedStart = Math.max(0, getCascadedStart(sequence, sequences));
-	if (!sequence.parent) {
-		return cascadedStart;
-	}
-
-	const parent = sequences.find((s) => s.id === sequence.parent);
-	if (!parent) {
-		throw new TypeError('Parent not found for sequence ' + sequence.id);
-	}
-
-	const timelineVisibleStart = getTimelineVisibleStart(parent, sequences);
-	return Math.max(timelineVisibleStart, cascadedStart);
-};
-
-export const getTimelineVisibleDuration = (
-	sequence: TSequence,
-	sequences: TSequence[],
-): number => {
-	const visibleDuration = sequence.duration + Math.min(sequence.from, 0);
-	if (!sequence.parent) {
-		return visibleDuration;
-	}
-
-	const parent = sequences.find((s) => s.id === sequence.parent);
-	if (!parent) {
-		throw new TypeError('Parent not found for sequence ' + sequence.id);
-	}
-
-	return Math.min(
-		visibleDuration,
-		getTimelineVisibleDuration(parent, sequences),
-	);
-};
+export {
+	getCascadedStart,
+	getCascadedStartWithTrim,
+	getTimelineVisibleDuration,
+	getTimelineVisibleStart,
+} from '@remotion/canvas';
