+++ packages/studio/src/helpers/get-timeline-sequence-sort-key.ts
@@ -1,49 +1,6 @@
-import type {
-	_InternalTypes,
-	SequencePropsSubscriptionKey,
-	TSequence,
-} from 'remotion';
-
-export type SequenceNodePathInfo = {
-	sequenceSubscriptionKey: SequencePropsSubscriptionKey;
-	auxiliaryKeys: string[];
-	index: number;
-	numberOfSequencesWithThisNodePath: number;
-	supportsEffects: boolean;
-};
-
-export type TimelineTrackData = {
-	sequence: TSequence;
-	connectedCompositions?: readonly _InternalTypes['AnyComposition'][];
-	depth: number;
-	nodePathInfo: SequenceNodePathInfo | null;
-	keyframeDisplayOffset: number;
-	sequenceFrameOffset: number;
-	cascadedStart: number;
-	localStart: number;
-};
-
-export type TimelineTrackWithOriginalTimings = TimelineTrackData & {
-	cascadedDuration: number;
-};
-
-export const getTimelineSequenceSortKey = (
-	track: TimelineTrackData,
-	tracks: TimelineTrackData[],
-	nonceRanks: Map<string, number>,
-): string => {
-	const rank = nonceRanks.get(track.sequence.id) ?? 0;
-	const id = String(rank).padStart(6, '0');
-	if (!track.sequence.parent) {
-		return id;
-	}
-
-	const parent = tracks.find((t) => t.sequence.id === track.sequence.parent);
-	if (!parent) {
-		// Due to effects and conditional `showInTimeline`, a parent
-		// may not exist in the `allTracks` array.
-		return id;
-	}
-
-	return `${getTimelineSequenceSortKey(parent, tracks, nonceRanks)}-${id}`;
-};
+export {getTimelineSequenceSortKey} from '@remotion/canvas';
+export type {
+	SequenceNodePathInfo,
+	TimelineTrackData,
+	TimelineTrackWithOriginalTimings,
+} from '@remotion/canvas';
