+++ packages/studio/src/helpers/timeline-node-path-key.ts
@@ -1,10 +1,7 @@
 import {stringifySequenceExpandedRowKey} from '@remotion/studio-shared';
-import type {SequencePropsSubscriptionKey} from 'remotion';
 import type {SequenceNodePathInfo} from './get-timeline-sequence-sort-key';
 
-export const timelineSequenceNodePathToKey = (
-	key: SequencePropsSubscriptionKey,
-): string => `${key.absolutePath}:${JSON.stringify(key.nodePath)}`;
+export {timelineSequenceNodePathToKey} from '@remotion/canvas';
 
 export const timelineNodePathInfoToKey = (info: SequenceNodePathInfo): string =>
 	[
